Add timeout option for MX DNS record lookup
Add optional second parameter `opts`, an object that can contain the
properties `timeout` and `checkMx`. The `checkMx` property of the `opts`
object overrides the `checkMx` function parameter. The `timeout`
property, in `ms` module format, specifies the timeout for the DNS MX
record lookup, the default is 10 seconds.

Add test, example and update the API documentation.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>